requirements.txt
aws-cdk.aws-ec2
aws-cdk.aws-ecs
aws-cdk.core
black
